copied gpc.py into gp_classification.py
